Change signed keys type to cert-equivalent
OpenSSH changes the basic key type to the certificate equivalent [1]. The
exact reason remains unclear, as it is never checked when reading a
certificate. However, it is possible this will be added in a future version.

To stay compatible with openssh as much as possible, we also change the signed
key type to the cert variant.

[1]: https://github.com/openssh/openssh-portable/blob/2dc328023f60212cd29504fc05d849133ae47355/ssh-keygen.c#L1828