Make Tests Consistent
When I was originally writing the Mozilla handler for this library, I
didn't implement hashing of the challenge because there is no way to
take in a challenge from a remote host. This is why @timweri had to
remove the hashing from the test in his earlier commits.

We now generate random data first, then hash it such that we can get the
preimage to test it properly. This also will allow us to extend the API
into the future to allow remote challenges for key generation
verification.